move vmStatus logic to utils/status/vm
