+++ src/components/VmConsoles/VmConsoles.js
@@ -4,7 +4,7 @@ import { get } from 'lodash';
 import { AccessConsoles, VncConsole } from '@patternfly/react-console';
 import { Button } from 'patternfly-react';
 
-import { isVmiRunning, isVmStarting } from '../VmStatus';
+import { isVmiRunning, isVmStarting } from '../../utils/status/vm';
 import { SerialConsoleConnector } from './SerialConsoleConnector';
 import { isWindows, isGuestAgentConnected } from '../../selectors';
 import { DEFAULT_RDP_PORT, TEMPLATE_VM_NAME_LABEL } from '../../constants';
