+++ src/utils/index.js
@@ -1,4 +1,5 @@
 export * from './mediaQuery';
+export * from './status';
 export * from './templates';
 export * from './utils';
 export * from './validations';
